src/samperium/requirements.txt
nextcord>=2.5.0,<3.0.0
python-dotenv==1.0.0